+++ transport/src/main/java/io/netty/channel/DefaultChannelHandlerInvoker.java
@@ -165,7 +165,7 @@ public class DefaultChannelHandlerInvoker implements ChannelHandlerInvoker {
         if (executor.inEventLoop()) {
             invokeChannelReadCompleteNow(ctx);
         } else {
-            DefaultChannelHandlerContext dctx = (DefaultChannelHandlerContext) ctx;
+            AbstractChannelHandlerContext dctx = (AbstractChannelHandlerContext) ctx;
             Runnable task = dctx.invokeChannelReadCompleteTask;
             if (task == null) {
                 dctx.invokeChannelReadCompleteTask = task = new Runnable() {
@@ -184,7 +184,7 @@ public class DefaultChannelHandlerInvoker implements ChannelHandlerInvoker {
         if (executor.inEventLoop()) {
             invokeChannelWritabilityChangedNow(ctx);
         } else {
-            DefaultChannelHandlerContext dctx = (DefaultChannelHandlerContext) ctx;
+            AbstractChannelHandlerContext dctx = (AbstractChannelHandlerContext) ctx;
             Runnable task = dctx.invokeChannelWritableStateChangedTask;
             if (task == null) {
                 dctx.invokeChannelWritableStateChangedTask = task = new Runnable() {
@@ -307,7 +307,7 @@ public class DefaultChannelHandlerInvoker implements ChannelHandlerInvoker {
         if (executor.inEventLoop()) {
             invokeReadNow(ctx);
         } else {
-            DefaultChannelHandlerContext dctx = (DefaultChannelHandlerContext) ctx;
+            AbstractChannelHandlerContext dctx = (AbstractChannelHandlerContext) ctx;
             Runnable task = dctx.invokeReadTask;
             if (task == null) {
                 dctx.invokeReadTask = task = new Runnable() {
@@ -353,7 +353,7 @@ public class DefaultChannelHandlerInvoker implements ChannelHandlerInvoker {
         if (executor.inEventLoop()) {
             invokeFlushNow(ctx);
         } else {
-            DefaultChannelHandlerContext dctx = (DefaultChannelHandlerContext) ctx;
+            AbstractChannelHandlerContext dctx = (AbstractChannelHandlerContext) ctx;
             Runnable task = dctx.invokeFlushTask;
             if (task == null) {
                 dctx.invokeFlushTask = task = new Runnable() {
